==> 7. Checking Warnings and False Convergences.Rmd <==
---
title: "7. Checking for Warnings and False Convergences"
author: "Bruce Tran"
date: "16 January 2018"
output: word_document
---

```{r}
##Analysis of Results:

setwd("F:\\University\\Honours\\Research Project\\Modelling\\Results\\Round 5")

##Fitted modelled flow (2000_2006)
sim6y1990.2   <- dir(path = ".\\Fitted_1990_1996_2\\", pattern = ".rdata")
sim6y1990.3   <- dir(path = ".\\Fitted_1990_1996_3\\", pattern = ".rdata")
sim10y1990.2  <- dir(path = ".\\Fitted_1990_1996_2\\", pattern = ".rdata")
sim10y1990.3  <- dir(path = ".\\Fitted_1990_1996_3\\", pattern = ".rdata")
sim6y2000.2   <- dir(path = ".\\Fitted_2000_2006_2\\", pattern = ".rdata")
sim6y2000.3   <- dir(path = ".\\Fitted_2000_2006_3\\", pattern = ".rdata")
sim10y2000.2  <- dir(path = ".\\Fitted_2000_2006_2\\", pattern = ".rdata")
sim10y2000.3  <- dir(path = ".\\Fitted_2000_2006_3\\", pattern = ".rdata")

obs_flow <- dir(path = "F:\\University\\Honours\\Research Project\\Modelling\\zoo SILO", pattern = ".Rdata")

##Making a table for the messages
Message <- data.frame(Number = rep(NA, length(obs_flow)), 
                      y6.1990.2  =  rep(NA, length(obs_flow)),
                      y6.1990.3  =  rep(NA, length(obs_flow)),
                      y10.1990.2 =  rep(NA, length(obs_flow)),
                      y10.1990.3 =  rep(NA, length(obs_flow)),
                      y6.2000.2  =  rep(NA, length(obs_flow)),
                      y6.2000.3  =  rep(NA, length(obs_flow)),
                      y10.2000.2 =  rep(NA, length(obs_flow)),
                      y10.2000.3 =  rep(NA, length(obs_flow)))

#Names
for (i in 1:length(obs_flow)) {
  load(paste("F:\\University\\Honours\\Research Project\\Modelling\\zoo SILO\\", obs_flow[i], sep = ""))
  
  ##Extracting number, name and Area for conversion
  river.number <- as.character(strsplit(obs_flow[i], ".Rdata"))
  
  Message[i, 1] <- river.number
}

##Making a list of functions
messages <-   function (x) {
    load(paste(dir()[1], "\\", x, sep = ""))
    
    message <- riverFit$fit.result$message
  }
message <- lapply(sim6y1990.2, FUN = messages)
Message$y6.1990.2 <- message

messages <-   function (x) {
  load(paste(dir()[2], "\\", x, sep = ""))
  
  message <- riverFit$fit.result$message
}
message <- lapply(sim6y1990.3, FUN = messages)
Message$y6.1990.3 <- message

messages <-   function (x) {
  load(paste(dir()[3], "\\", x, sep = ""))
  
  message <- riverFit$fit.result$message
}
message <- lapply(sim10y1990.2, FUN = messages)
Message$y10.1990.2 <- message

messages <-   function (x) {
  load(paste(dir()[4], "\\", x, sep = ""))
  
  message <- riverFit$fit.result$message
}
message <- lapply(sim10y1990.3, FUN = messages)
Message$y10.1990.3 <- message

messages <-   function (x) {
  load(paste(dir()[5], "\\", x, sep = ""))
  
  message <- riverFit$fit.result$message
}
message <- lapply(sim6y1990.2, FUN = messages)
Message$y6.2000.2 <- message

messages <-   function (x) {
  load(paste(dir()[6], "\\", x, sep = ""))
  
  message <- riverFit$fit.result$message
}
message <- lapply(sim6y1990.3, FUN = messages)
Message$y6.2000.3 <- message

messages <-   function (x) {
  load(paste(dir()[7], "\\", x, sep = ""))
  
  message <- riverFit$fit.result$message
}
message<- lapply(sim10y1990.2, FUN = messages)
Message$y10.2000.2 <- message

messages <-   function (x) {
  load(paste(dir()[8], "\\", x, sep = ""))
  
  message <- riverFit$fit.result$message
}
message <- lapply(sim10y1990.3, FUN = messages)
Message$y10.2000.3 <- message

##Taking care of the "unimplemented type 'list' in 'EncodeElement" error
for (i in 1:8) {
  Message[,i+1] <- vapply(Message[,i+1], paste, collapse = ", ", character(1L))
}

write.csv(Message, file = "E:\\University\\Honours\\Data Access\\Modelling\\Results\\Convergence Checks.csv", row.names = FALSE)
```

This is to get a list of catchments that could not converge during the analysis. This was done for all simulations in the model. 
In hindsight, if I had chosen a 6 or 10 year period first, it would have made things less time-consuming. 
The warning messages for each simulation was noted down, and based on these results, either "Fitted_2" or "Fitted_3" was chosen, based on which had the least false convergences.
